Refactor: extract favourites page module into web/js/pages/favourites.js (#80)
